v3 snapshot
	modified:   adc.c
	modified:   adc.h
	modified:   flashlight.c
	modified:   led.h
	modified:   routines.c
	modified:   routines.h

+++ routines.h
@@ -13,4 +13,4 @@ void start_timer1_ctc(char mode,unsigned int value);
 void init_mpu(void);
 
 void set_brightness(unsigned short mode);
-void indicator(unsigned short adc);
+void indicator(unsigned int adc);
